+++ data-raw/V_DES_COD.R
@@ -4,30 +4,31 @@ library(data.table)
 library(askpass)
 library(inesss)
 library(lubridate)
-# conn <- sql_connexion(askpass("User"))
+# conn <- SQL_connexion(askpass("User"))
 
 fct <- function(need_conn = FALSE) {
 
+  ### Effectuer la connexion à Teradata
   if (need_conn) {
     conn <- sql_connexion(askpass("User"))
   }
 
+  ### Extraction de la table
   query <-
     "select CODE_VAL_COD as CODE,
     	CODE_NOM_COD as TYPE_CODE,
     	CODE_DES as CODE_DESC,
     	CODE_DD_DES_COD as DATE_DEBUT,
     	CODE_DF_DES_COD as DATE_FIN
-    from PROD.V_DES_COD
-    order by CODE_NOM_COD asc, CODE_VAL_COD asc;"
-
+    from PROD.V_DES_COD;"
   DT <- as.data.table(dbGetQuery(conn, query))
 
-  # Modifier les dates aberrantes > date actuelle
-  DT[CODE_DF_DES_COD > Sys.Date(), CODE_DF_DES_COD := as_date(paste0(year(Sys.Date()),"12-31"))]
+  ### Arrangement des données
+  DT[, `:=` (DATE_DEBUT = as_date(DATE_DEBUT),  # s'assurer que c'est au format date
+             DATE_FIN = as_date(DATE_FIN))]
+  setkey(DT, TYPE_CODE, CODE, DATE_DEBUT)  # tri
 
-  setkey(DT, TYPE_CODE, CODE)
-  attr(DT, "Date") <- Sys.Date()
+  attr(DT, "Date") <- Sys.Date()  # date de création
   return(DT)
 
 }